Cleared up simple compiler warnings on i386_linux24

+++ StRoot/StHbtMaker/Infrastructure/StHbtAihongPid.h
@@ -1,6 +1,6 @@
 /***************************************************************************
  *
- * $Id: StHbtAihongPid.h,v 1.2 2002/01/14 16:10:57 laue Exp $
+ * $Id: StHbtAihongPid.h,v 1.3 2003/01/31 19:44:00 magestro Exp $
  *
  * Author: Frank Laue, Ohio State, laue@bnl.gov
  ***************************************************************************
@@ -27,9 +27,9 @@ class StHbtAihongPid  {
   static StHbtAihongPid* Instance(const char* fileName="PIDTable.root");
   void setPidTable(const char* fileName);
   void updateEvent(int refMult);
-  void updateTrack(int charge, double p, double eta, double nHitsDedx, double dedx);
-  StuProbabilityPidAlgorithm* aihongPid(int refMult, int charge, double p, double eta, double nHitsDedx, double dedx);
-  StuProbabilityPidAlgorithm* aihongPid(int charge, double p, double eta, double nHitsDedx, double dedx);
+  void updateTrack(int charge, double p, double eta, int nHitsDedx, double dedx);
+  StuProbabilityPidAlgorithm* aihongPid(int refMult, int charge, double p, double eta, int nHitsDedx, double dedx);
+  StuProbabilityPidAlgorithm* aihongPid(int charge, double p, double eta, int nHitsDedx, double dedx);
   StuProbabilityPidAlgorithm* aihongPid();
   
   friend class nobody;
